backend/requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.2
factory-boy>=3.3
faker>=22.0
coverage>=7.4
